+++ src/_gettsim/transfers/arbeitsl_geld_2/kindergelduebertrag.py
@@ -101,10 +101,13 @@ def kindergeld_zur_bedarfsdeckung_m(
     )
 
 
-def _diff_kindergeld_kindbedarf_m(
-    _arbeitsl_geld_2_eink_ohne_kindergeldübertrag_m_bg: float,
+def _diff_kindergeld_kindbedarf_m(  # noqa: PLR0913
     arbeitsl_geld_2_regelbedarf_m_bg: float,
+    _arbeitsl_geld_2_nettoeink_ohne_transfers_m: float,
+    wohngeld_nach_vermög_check_m_bg: float,
     kindergeld_zur_bedarfsdeckung_m: float,
+    kind_unterh_erhalt_m: float,
+    unterhaltsvors_m: float,
     eigenbedarf_gedeckt: bool,
 ) -> float:
     """Kindergeld that is used to cover the needs (SGB II) of the parent.
@@ -118,12 +121,18 @@ def _diff_kindergeld_kindbedarf_m(
 
     Parameters
     ----------
-    _arbeitsl_geld_2_eink_ohne_kindergeldübertrag_m_bg
-        See :func:`_arbeitsl_geld_2_eink_ohne_kindergeldübertrag_m_bg`.
     arbeitsl_geld_2_regelbedarf_m_bg
         See :func:`arbeitsl_geld_2_regelbedarf_m_bg`.
+    _arbeitsl_geld_2_nettoeink_ohne_transfers_m
+        See :func:`_arbeitsl_geld_2
+    wohngeld_nach_vermög_check_m_bg
+        See :func:`wohngeld_nach_vermög_check_m_bg`.
     kindergeld_zur_bedarfsdeckung_m
         See :func:`kindergeld_zur_bedarfsdeckung_m`.
+    kind_unterh_erhalt_m
+        See :func:`kind_unterh_erhalt_m`.
+    unterhaltsvors_m
+        See :func:`unterhaltsvors_m`.
     eigenbedarf_gedeckt
         See :func:`eigenbedarf_gedeckt`.
 
@@ -136,12 +145,13 @@ def _diff_kindergeld_kindbedarf_m(
     # Kindergeld would be counted twice as income of the Bedarfsgemeinschaft (one time
     # the full amount for the child and one time the Kindergeldübertrag for the parent -
     # because the child doesn't drop out of Bedarfsgemeinschaft endogenously).
-    # https://github.com/iza-institute-of-labor-economics/gettsim/issues/622
-    # TODO (@MImmesberger): Consider Kinderwohngeld in the Fehlbetrag calculation.
-    # https://github.com/iza-institute-of-labor-economics/gettsim/issues/750
+    # https://github.com/iza-institute-of-labor-economics/gettsim/issues/758
     fehlbetrag = max(
         arbeitsl_geld_2_regelbedarf_m_bg
-        - _arbeitsl_geld_2_eink_ohne_kindergeldübertrag_m_bg,
+        - wohngeld_nach_vermög_check_m_bg
+        - _arbeitsl_geld_2_nettoeink_ohne_transfers_m
+        - kind_unterh_erhalt_m
+        - unterhaltsvors_m,
         0.0,
     )
     # Bedarf not covered or same Bedarfsgemeinschaft as parents